Fixing missing particle forces with default init
