+++ src/wash/particle.cpp
@@ -1,0 +1,16 @@
+#include "wash.hpp"
+
+namespace wash {
+    void Particle::initialise_particle_forces() {
+        this->force_scalars = std::unordered_map<std::string, double>({});
+        this->force_vectors = std::unordered_map<std::string, wash::SimulationVecT>({});
+
+        for (auto& force : get_forces_scalar()) {
+            this->force_scalars[force] = 0.0;
+        }
+
+        for (auto& force : get_forces_vector()) {
+            this->force_vectors[force] = SimulationVecT{};
+        }
+    }
+}
