+++ src/wash/wash.cpp
@@ -117,20 +117,4 @@ namespace wash {
         auto pos = p.get_pos() - q.get_pos();
         return std::sqrt(pos.magnitude());
     }
-    
-    template <typename... Args>
-    std::string string_format(const std::string& format, Args... args) {
-        int size_s = std::snprintf(nullptr, 0, format.c_str(), args...) + 1;  // Extra space for '\0'
-        if (size_s <= 0) {
-            throw std::runtime_error("Error during formatting.");
-        }
-        auto size = static_cast<size_t>(size_s);
-        std::unique_ptr<char[]> buf(new char[size]);
-        std::snprintf(buf.get(), size, format.c_str(), args...);
-        return std::string(buf.get(), buf.get() + size - 1);  // We don't want the '\0' inside
-    }
-
-    template <typename T> int sgn(T val) {
-        return (T(0) < val) - (val < T(0));
-    }
 }
